feat: Add 'ANY' pseudo-record type
Adds a pseudo record type "ANY" that queries the given domain for all known record types.

This is implemented by adding an `all_record_types()` function that returns a vector of all `RecordType` variants. The command-line option parsing logic is updated to check for "ANY" and, if present, populate the record_types field with the list of all record types.

The test suite has been updated to reflect this new functionality.